!add_event can now be used in private channels.
Functionality was added to add_event() and check_event() that makes it possible to use the !add_event command in private chats. This is done by saving the message.author.id to the events.csv file instead of the message.channel.id if the channel it was sent from is a dm channel. This makes it possible to send a dm to the author id in the check_events function.